explain.js video recording: prefer current tab:
In current versions of chrome, the current tab did not show up at all
in the tab list on my machine (MacOS).
Additionally, it was always annoying to search through the list if many
tabs were opened.
A recent chrome feature makes this selection much easier by setting
`preferCurrentTab`.

https://chromestatus.com/feature/5045313003847680

According to
https://groups.google.com/a/chromium.org/g/blink-dev/c/YoefXLTQsw0/m/NVoV6NcIAgAJ ,
this does not cause issues in Gecko and Webkit (however I did not try).