Make scrolling in handout mode more reliable.
Change the API of updateCurrentSlideMark of the menu plugin and
requestMenuContext of the feedback plugin to also accept
the slide as a parameter before trying to fetch the current
slide value from reveal to make it work with the handout mode.

Now most functions trying to identify slides use their slide-ids
instead of asking reveal for indices and using them as there
seems to be no functional difference between just jumping to a slide
with a certain id and telling reveal to maneauver to a certain index.