Chrome pdf export protocol description
As reveal.js has an event that fires when it finished loading everything
needed for the pdf export one can simply listen and wait for this event
to occur before exporting. This can be done via chromes remote-debugging
api with a few messages. This commit adds a description for what
messages are needed, so I can later on implement this behaviour into
decker. This would also mean, that only one instance of headless chrome
is needed for all pdf pages to be exported.